Pass through tungstenite errors from send functions
Make the send_string and send_bytes functions pass through errors from
tungstenite. This allows the caller to inspect those errors, and to use
those errors as an implementation of std::error::Error.

The stream for incoming messages already provides access to the
tungstenite error; this makes the functions for outgoing messages do the
same.